change position mapping to long mysteriously fixes tests

+++ cms/tests/test_views.py
@@ -41,7 +41,7 @@ class TestViews(UnicoreTestCase):
                     'index': 'not_analyzed',
                 },
                 'position': {
-                    'type': 'integer'
+                    'type': 'long'
                 }
             }
         })
